change util.misc
